Create dedup
dedup is a bash script that removes duplicate reads from paired-end libraries generated with random base molecular indices (also known as unique molecular identifiers) on both ends.
Default parameters are hard-coded at the beginning of the script. I recommend changing Threads and BufferSize to match your environment (for example, Threads=80 and BufferSize=5G is appropriate for an 88-core workstation with 512 GB of RAM).